refactor(lsp): split service responsibilities
