--- internal/lsp/service.go
@@ -1,2165 +1,0 @@
-package lsp
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"log"
-	"net/url"
-	"os"
-	"path/filepath"
-	"runtime"
-	"slices"
-	"strings"
-	"sync"
-	"time"
-	"unicode"
-
-	"github.com/microsoft/typescript-go/shim/ast"
-	"github.com/microsoft/typescript-go/shim/bundled"
-	"github.com/microsoft/typescript-go/shim/compiler"
-	"github.com/microsoft/typescript-go/shim/core"
-	"github.com/microsoft/typescript-go/shim/lsp/lsproto"
-	"github.com/microsoft/typescript-go/shim/project"
-	"github.com/microsoft/typescript-go/shim/project/logging"
-	"github.com/microsoft/typescript-go/shim/scanner"
-	"github.com/microsoft/typescript-go/shim/tsoptions"
-	"github.com/microsoft/typescript-go/shim/tspath"
-	"github.com/microsoft/typescript-go/shim/vfs"
-	"github.com/microsoft/typescript-go/shim/vfs/cachedvfs"
-
-	"github.com/web-infra-dev/rslint/internal/config"
-	"github.com/web-infra-dev/rslint/internal/config/discovery"
-	"github.com/web-infra-dev/rslint/internal/config/target"
-	"github.com/web-infra-dev/rslint/internal/linter"
-	lintprogram "github.com/web-infra-dev/rslint/internal/program"
-	"github.com/web-infra-dev/rslint/internal/rule"
-	"github.com/web-infra-dev/rslint/internal/rules"
-	"github.com/web-infra-dev/rslint/internal/utils"
-)
-
-const codeActionKindSourceFixAllRslint = lsproto.CodeActionKind("source.fixAll.rslint")
-const gitignoreWatcherID project.WatcherID = "rslint-gitignore-policy"
-const ancestorJSConfigWatcherID project.WatcherID = "rslint-ancestor-js-config"
-
-type lintPassResult struct {
-	Diagnostics     []rule.RuleDiagnostic
-	HasSyntaxErrors bool
-}
-
-// documentLintSnapshot freezes every config-sensitive fact for one editor
-// operation. Owner selection, project selection, Go rules, plugin rules,
-// and fixes must all consume this same target/config pair; none may rediscover
-// the target identity from the filesystem mid-operation.
-type documentLintSnapshot struct {
-	target                target.File
-	config                config.RslintConfig
-	resolvedConfig        config.ResolvedFileConfig
-	pathSpaces            *config.PathSpaceSnapshot
-	ruleCatalog           *rule.Catalog
-	configResolved        bool
-	typeScriptConfigPaths []string
-	usesJavaScriptConfig  bool
-	unavailable           bool
-}
-
-func (s *Server) currentRuleCatalog() *rule.Catalog {
-	if s == nil || s.ruleCatalog == nil {
-		panic("LSP rule catalog is not initialized")
-	}
-	return s.ruleCatalog
-}
-
-func resolveDocumentLintSnapshotConfig(
-	snapshot documentLintSnapshot,
-	fs vfs.FS,
-) documentLintSnapshot {
-	if snapshot.ruleCatalog == nil {
-		panic("document lint snapshot requires a rule catalog")
-	}
-	if snapshot.pathSpaces == nil {
-		panic("document lint snapshot requires a path-space snapshot")
-	}
-	resolver, err := config.NewFileConfigResolverWithPathSpaces(
-		snapshot.config,
-		snapshot.target.ConfigDirectory,
-		fs,
-		snapshot.pathSpaces,
-		snapshot.ruleCatalog,
-		snapshot.usesJavaScriptConfig,
-	)
-	if err != nil {
-		panic(err)
-	}
-	snapshot.resolvedConfig = resolver.ResolveTarget(snapshot.target.Identity())
-	snapshot.configResolved = true
-	return snapshot
-}
-
-// ruleFixToTextEdit converts a rule fix into an LSP TextEdit using the
-// source file's line map for position encoding.
-func ruleFixToTextEdit(sourceFile ast.SourceFileLike, fix rule.RuleFix) *lsproto.TextEdit {
-	startLine, startChar := scanner.GetECMALineAndUTF16CharacterOfPosition(sourceFile, fix.Range.Pos())
-	endLine, endChar := scanner.GetECMALineAndUTF16CharacterOfPosition(sourceFile, fix.Range.End())
-	return &lsproto.TextEdit{
-		Range: lsproto.Range{
-			Start: lsproto.Position{Line: uint32(startLine), Character: uint32(startChar)},
-			End:   lsproto.Position{Line: uint32(endLine), Character: uint32(endChar)},
-		},
-		NewText: fix.Text,
-	}
-}
-
-func (s *Server) handleInitialize(ctx context.Context, params *lsproto.InitializeParams) (lsproto.InitializeResponse, error) {
-	log.Printf("handle initialize with pid: %d\n", os.Getpid())
-	if s.initializeParams != nil {
-		return nil, lsproto.ErrorCodeInvalidRequest
-	}
-
-	s.initializeParams = params
-
-	// rslint diagnostics and edits use VS Code's native UTF-16 coordinates.
-	// UTF-16 is mandatory for LSP clients, so selecting it unconditionally keeps
-	// incoming incremental edits and outgoing ranges on one encoding contract.
-	s.positionEncoding = lsproto.PositionEncodingKindUTF16
-
-	response := &lsproto.InitializeResult{
-		ServerInfo: &lsproto.ServerInfo{
-			Name:    "typescript-go",
-			Version: ptrTo(core.Version()),
-		},
-		Capabilities: &lsproto.ServerCapabilities{
-			PositionEncoding: ptrTo(s.positionEncoding),
-			TextDocumentSync: &lsproto.TextDocumentSyncOptionsOrKind{
-				Options: &lsproto.TextDocumentSyncOptions{
-					OpenClose: ptrTo(true),
-					Change:    ptrTo(lsproto.TextDocumentSyncKindIncremental),
-					Save: &lsproto.BooleanOrSaveOptions{
-						SaveOptions: &lsproto.SaveOptions{
-							IncludeText: ptrTo(true),
-						},
-					},
-				},
-			},
-			CodeActionProvider: &lsproto.BooleanOrCodeActionOptions{
-				CodeActionOptions: &lsproto.CodeActionOptions{
-					CodeActionKinds: &[]lsproto.CodeActionKind{
-						lsproto.CodeActionKindQuickFix,
-						lsproto.CodeActionKindSourceFixAll,
-						codeActionKindSourceFixAllRslint,
-					},
-				},
-			},
-		},
-	}
-
-	return response, nil
-}
-func (s *Server) handleInitialized(ctx context.Context, params *lsproto.InitializedParams) error {
-	// Enable file watching if the client supports dynamic registration of
-	// didChangeWatchedFiles. This allows Session to register tsconfig watchers
-	// and call RefreshDiagnostics when project state changes.
-	if s.initializeParams.Capabilities != nil &&
-		s.initializeParams.Capabilities.Workspace != nil &&
-		s.initializeParams.Capabilities.Workspace.DidChangeWatchedFiles != nil &&
-		ptrIsTrue(s.initializeParams.Capabilities.Workspace.DidChangeWatchedFiles.DynamicRegistration) {
-		s.watchEnabled = true
-	}
-	if s.watchEnabled && s.outgoingQueue != nil {
-		relativePatterns := ptrIsTrue(
-			s.initializeParams.Capabilities.Workspace.DidChangeWatchedFiles.RelativePatternSupport,
-		)
-		gitignoreWatchers := gitignoreFileWatchers(s.cwd, relativePatterns)
-		ancestorConfigWatchers := ancestorJSConfigFileWatchers(s.cwd, relativePatterns)
-		go func() {
-			if err := s.WatchFiles(s.backgroundCtx, gitignoreWatcherID, gitignoreWatchers); err != nil {
-				if s.backgroundCtx.Err() == nil {
-					log.Printf("[rslint] Failed to register .gitignore watchers: %v", err)
-				}
-			}
-			if len(ancestorConfigWatchers) == 0 {
-				return
-			}
-			if err := s.WatchFiles(s.backgroundCtx, ancestorJSConfigWatcherID, ancestorConfigWatchers); err != nil {
-				if s.backgroundCtx.Err() == nil {
-					log.Printf("[rslint] Failed to register ancestor JS config watchers: %v", err)
-				}
-			}
-		}()
-	}
-
-	s.session = project.NewSession(&project.SessionInit{
-		BackgroundCtx: s.backgroundCtx,
-		Options: &project.SessionOptions{
-			CurrentDirectory:   s.cwd,
-			DefaultLibraryPath: s.defaultLibraryPath,
-			TypingsLocation:    s.typingsLocation,
-			PositionEncoding:   s.positionEncoding,
-			WatchEnabled:       s.watchEnabled,
-		},
-		FS:         s.fs,
-		Client:     s,
-		Logger:     logging.NewLogger(io.Discard),
-		ParseCache: s.parseCache,
-	})
-	if s.watchEnabled && s.outgoingQueue != nil {
-		s.lintPrograms = newLintProgramStore(s)
-	}
-
-	// Load the JSON config used before the first JS/TS catalog transaction and
-	// as fallback for files outside every discovered JS/TS config boundary.
-	rslintConfigPath, configFound := findRslintConfig(s.fs, s.cwd)
-	if configFound {
-		s.rslintConfigPath = rslintConfigPath
-		if err := s.reloadConfig(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func gitignoreFileWatchers(cwd string, relativePatternSupport bool) []*lsproto.FileSystemWatcher {
-	workspaceRoot := filepath.Clean(cwd)
-	watchers := []*lsproto.FileSystemWatcher{
-		fileSystemWatcher(workspaceRoot, "**/.gitignore", relativePatternSupport),
-	}
-	// Automatic discovery may select a config above the workspace. Exact
-	// watchers on the strict lexical ancestors cover every possible source from
-	// that config directory down to cwd without recursively watching siblings.
-	for current := filepath.Dir(workspaceRoot); current != workspaceRoot; {
-		watchers = append(watchers, fileSystemWatcher(current, ".gitignore", relativePatternSupport))
-		parent := filepath.Dir(current)
-		if parent == current {
-			break
-		}
-		current = parent
-	}
-	return watchers
-}
-
-// ancestorJSConfigFileWatchers covers the strict lexical ancestors that Go's
-// automatic config discovery searches before walking the workspace. The
-// extension already owns a workspace-scoped RelativePattern watcher, so the
-// workspace itself is deliberately excluded to avoid duplicate refreshes.
-// Register every filename separately: creating a higher-priority sibling in an
-// ancestor directory can change the selected config even when another config
-// basename already exists there.
-func ancestorJSConfigFileWatchers(cwd string, relativePatternSupport bool) []*lsproto.FileSystemWatcher {
-	workspaceRoot := filepath.Clean(cwd)
-	watchers := make([]*lsproto.FileSystemWatcher, 0)
-	for current := filepath.Dir(workspaceRoot); current != workspaceRoot; {
-		for _, configName := range discovery.AutoJSConfigFileNames {
-			watchers = append(watchers, fileSystemWatcher(current, configName, relativePatternSupport))
-		}
-		parent := filepath.Dir(current)
-		if parent == current {
-			break
-		}
-		current = parent
-	}
-	return watchers
-}
-
-func fileSystemWatcher(baseDir string, pattern string, relativePatternSupport bool) *lsproto.FileSystemWatcher {
-	if relativePatternSupport {
-		uri := fileURIFromPath(baseDir)
-		return &lsproto.FileSystemWatcher{
-			GlobPattern: lsproto.PatternOrRelativePattern{
-				RelativePattern: &lsproto.RelativePattern{
-					BaseUri: lsproto.WorkspaceFolderOrURI{URI: &uri},
-					Pattern: pattern,
-				},
-			},
-		}
-	}
-	absolute := filepath.ToSlash(filepath.Join(baseDir, pattern))
-	return &lsproto.FileSystemWatcher{
-		GlobPattern: lsproto.PatternOrRelativePattern{Pattern: &absolute},
-	}
-}
-
-func fileURIFromPath(filePath string) lsproto.URI {
-	uriPath := filepath.ToSlash(filePath)
-	if len(uriPath) >= 2 && uriPath[1] == ':' {
-		uriPath = "/" + uriPath
-	}
-	return lsproto.URI((&url.URL{Scheme: "file", Path: uriPath}).String())
-}
-
-// reloadConfig loads (or reloads) the rslint JSON configuration from s.rslintConfigPath.
-// The LSP reuses projects already loaded by project service and builds a
-// session-external ts-go Program for a declared custom project. Resolving
-// project paths here preserves declaration order and ensures type-aware rules
-// run only when the governing config's first containing project supplies type
-// information.
-func (s *Server) reloadConfig() error {
-	loader := config.NewConfigLoader(s.fs, s.cwd, rules.All())
-	rslintConfig, _, err := loader.LoadRslintConfig(s.rslintConfigPath)
-	if err != nil {
-		return fmt.Errorf("could not load rslint config: %w", err)
-	}
-	paths, err := s.resolveTsConfigPaths(rslintConfig, s.cwd)
-	if err != nil {
-		return fmt.Errorf("could not resolve tsconfig paths for %q: %w", s.rslintConfigPath, err)
-	}
-	ownerIndex, fileConfigResolver, err := prepareJSONConfigEvaluation(
-		rslintConfig,
-		s.cwd,
-		s.fs,
-	)
-	if err != nil {
-		return fmt.Errorf("could not prepare JSON config evaluation: %w", err)
-	}
-	s.jsonConfig = rslintConfig
-	s.jsonConfigOwnerIndex = ownerIndex
-	s.jsonFileConfigResolver = fileConfigResolver
-	s.tsConfigPaths = paths
-	s.invalidateLintProjectCaches()
-	return nil
-}
-
-func prepareJSONConfigEvaluation(
-	entries config.RslintConfig,
-	configDirectory string,
-	fsys vfs.FS,
-) (*target.OwnerIndex, *config.FileConfigResolver, error) {
-	configDirectory = tspath.NormalizePath(configDirectory)
-	ownerIndex := target.NewOwnerIndex(
-		map[string]config.RslintConfig{configDirectory: entries},
-		fsys,
-	)
-	resolver, err := config.NewFileConfigResolverWithPathSpaces(
-		entries,
-		configDirectory,
-		fsys,
-		ownerIndex.PathSpaces(),
-		rules.All(),
-		false,
-	)
-	if err != nil {
-		return nil, nil, err
-	}
-	return ownerIndex, resolver, nil
-}
-
-func (s *Server) invalidateLintProjectCaches() {
-	if s.lintPrograms != nil {
-		s.lintPrograms.Invalidate()
-	}
-	if s.lintSessionRoots != nil {
-		s.lintSessionRoots.Invalidate()
-	}
-}
-
-func validateRuleOptionsForConfig(
-	entries config.RslintConfig,
-	configDirectory string,
-	catalog *rule.Catalog,
-) (config.RslintConfig, error) {
-	normalized, optionsErrs := config.ValidateRuleOptions(entries, catalog)
-	if len(optionsErrs) == 0 {
-		return normalized, nil
-	}
-	msgs := make([]string, len(optionsErrs))
-	for i, optionsErr := range optionsErrs {
-		msgs[i] = optionsErr.Error()
-	}
-	return nil, fmt.Errorf("invalid rule options for %q:\n%s", configDirectory, strings.Join(msgs, "\n"))
-}
-
-// handleDidChangeWatchedFiles handles file change notifications from the client.
-func (s *Server) handleDidChangeWatchedFiles(ctx context.Context, params *lsproto.DidChangeWatchedFilesParams) error {
-	if params == nil {
-		return nil
-	}
-
-	if s.lintPrograms != nil &&
-		s.lintPrograms.DidChangeWatchedFiles(params.Changes) {
-		s.invalidateOpenDocumentDiagnostics()
-		_ = s.RefreshDiagnostics(ctx)
-	}
-
-	// Preserve Session's original watched-file input. It owns configured
-	// project identity and may need a disk event while an overlay is open.
-	if s.session != nil {
-		s.session.DidChangeWatchedFiles(ctx, params.Changes)
-	}
-	// Check for config file changes that affect rslint.
-	needsConfigReload := false
-	needsTypeInfoRebuild := false
-	needsIgnoreRefresh := false
-	needsAncestorJSConfigRefresh := false
-	for _, change := range params.Changes {
-		uri := string(change.Uri)
-		if isRslintConfigURI(uri) {
-			needsConfigReload = true
-		}
-		if isTsConfigURI(uri) {
-			needsTypeInfoRebuild = true
-		}
-		if isGitignoreURI(uri) {
-			needsIgnoreRefresh = true
-		}
-		if isStrictAncestorAutoJSConfigPath(uriToPath(change.Uri), s.cwd, s.fs) {
-			needsAncestorJSConfigRefresh = true
-		}
-	}
-	needsAutomaticAncestorRefresh := needsAncestorJSConfigRefresh && s.configRefreshConfigPath == ""
-	if (needsIgnoreRefresh || needsAutomaticAncestorRefresh) && s.configDiscoveryActive {
-		// didChangeWatchedFiles and configRefresh are both blocking methods, so
-		// this direct call stays on the server's serialized dispatch loop and
-		// cannot race an extension-initiated transaction. JSON fallback is part
-		// of the candidate snapshot: never reload it directly while discovery is
-		// active, otherwise a later JS activation failure could leave half of a
-		// rejected generation live.
-		reason := "gitignore-change"
-		if needsAutomaticAncestorRefresh {
-			reason = "config-change"
-		}
-		_, err := s.refreshConfig(ctx)
-		if err == nil {
-			return nil
-		}
-		// Discovery/activation failure preserves the complete last-good
-		// generation, including its .gitignore view. Recompute diagnostics from
-		// that committed view after an ignore event so invalidated editor
-		// results are republished without leaking the rejected filesystem state. A
-		// config-only failure has no independently live state to invalidate.
-		log.Printf("[rslint] Failed to refresh config catalog after watched %s: %v", reason, err)
-		if needsIgnoreRefresh {
-			s.invalidateOpenDocumentDiagnostics()
-			return s.RefreshDiagnostics(ctx)
-		}
-		return nil
-	}
-	if s.configDiscoveryActive {
-		// The extension's direct workspace watcher is the sole owner for
-		// workspace/descendant JS configs and JSON fallback. tsgo can also report
-		// those paths through its recursive project watcher; treating that report
-		// as a second refresh would evaluate every fresh module twice. Go-owned
-		// didChange handling above is intentionally limited to .gitignore and
-		// strict-ancestor JS configs, which the extension watcher cannot cover.
-		needsConfigReload = false
-	}
-	if needsConfigReload {
-		s.reloadConfigAndRelint()
-		if needsIgnoreRefresh {
-			s.invalidateOpenDocumentDiagnostics()
-			return s.RefreshDiagnostics(ctx)
-		}
-		return nil
-	}
-	if needsTypeInfoRebuild {
-		// tsconfig changed — rebuild tsConfigPaths so type-aware rule filtering
-		// stays in sync. Session already handles the project state update and
-		// triggers RefreshDiagnostics for relinting.
-		if err := s.rebuildTsConfigPaths(); err != nil {
-			log.Printf("[rslint] Failed to rebuild tsconfig paths: %v", err)
-		}
-	}
-	if needsIgnoreRefresh {
-		s.invalidateOpenDocumentDiagnostics()
-		return s.RefreshDiagnostics(ctx)
-	}
-
-	return nil
-}
-
-// isRslintConfigURI returns true if the URI points to an rslint config file.
-func isRslintConfigURI(uri string) bool {
-	return strings.HasSuffix(uri, "/rslint.json") || strings.HasSuffix(uri, "/rslint.jsonc")
-}
-
-func isGitignoreURI(uri string) bool {
-	idx := strings.LastIndex(uri, "/")
-	return idx >= 0 && strings.EqualFold(uri[idx+1:], ".gitignore")
-}
-
-func isStrictAncestorAutoJSConfigPath(filePath string, cwd string, fsys vfs.FS) bool {
-	if filePath == "" || cwd == "" || fsys == nil {
-		return false
-	}
-	caseSensitive := fsys.UseCaseSensitiveFileNames()
-	baseName := tspath.GetBaseFileName(tspath.NormalizePath(filePath))
-	isAutoConfig := false
-	for _, configName := range discovery.AutoJSConfigFileNames {
-		if pathStringsEqual(baseName, configName, caseSensitive) {
-			isAutoConfig = true
-			break
-		}
-	}
-	if !isAutoConfig {
-		return false
-	}
-	directory := tspath.GetDirectoryPath(tspath.NormalizePath(filePath))
-	workspace := tspath.NormalizePath(cwd)
-	return !pathStringsEqual(directory, workspace, caseSensitive) &&
-		tspath.StartsWithDirectory(workspace, directory, caseSensitive)
-}
-
-func pathStringsEqual(left string, right string, caseSensitive bool) bool {
-	if caseSensitive {
-		return left == right
-	}
-	return strings.EqualFold(left, right)
-}
-
-func (s *Server) invalidateOpenDocumentDiagnostics() {
-	for uri := range s.documents {
-		s.docGeneration[uri]++
-		s.cancelInflightPluginDispatch(uri)
-		delete(s.diagnostics, uri)
-	}
-}
-
-// isTsConfigURI returns true if the URI points to a tsconfig/jsconfig file,
-// including variants like tsconfig.build.json, tsconfig.app.json, etc.
-func isTsConfigURI(uri string) bool {
-	idx := strings.LastIndex(uri, "/")
-	if idx < 0 {
-		return false
-	}
-	name := uri[idx+1:]
-	return (strings.HasPrefix(name, "tsconfig") || strings.HasPrefix(name, "jsconfig")) &&
-		strings.HasSuffix(name, ".json")
-}
-
-// resolveTsConfigPaths resolves parserOptions.project from a config while
-// preserving each declared path. TypeScript resolves relative includes from
-// that lexical location, so a symlinked tsconfig is not interchangeable with
-// its physical target.
-func (s *Server) resolveTsConfigPaths(cfg config.RslintConfig, cwd string) ([]string, error) {
-	return resolveTsConfigPathsWithFS(cfg, cwd, s.fs)
-}
-
-// rebuildTsConfigPaths resolves parserOptions.project from the current config.
-// Called when a tsconfig or rslint config changes so that type-aware rule
-// filtering stays in sync.
-//
-// For JS/TS configs we resolve per-config directory into tsConfigPathsByConfig.
-// A config whose parserOptions.project is empty and has no auto-detected
-// tsconfig resolves to nil. Files governed by that config have no type info,
-// without affecting files governed by other configs. A nested template or
-// fixture config without a tsconfig must not change sibling config behavior.
-func (s *Server) rebuildTsConfigPaths() error {
-	var tsConfigPaths []string
-	if s.rslintConfigPath != "" {
-		var err error
-		tsConfigPaths, err = s.resolveTsConfigPaths(s.jsonConfig, s.cwd)
-		if err != nil {
-			return fmt.Errorf("resolve tsconfig paths for %q: %w", s.rslintConfigPath, err)
-		}
-	}
-
-	var byConfig map[string][]string
-	if len(s.jsConfigs) > 0 {
-		byConfig = make(map[string][]string, len(s.jsConfigs))
-		for dir, entries := range s.jsConfigs {
-			paths, err := s.resolveTsConfigPaths(entries, dir)
-			if err != nil {
-				return fmt.Errorf("resolve tsconfig paths for %q: %w", dir, err)
-			}
-			byConfig[dir] = paths
-		}
-	}
-
-	s.tsConfigPaths = tsConfigPaths
-	s.tsConfigPathsByConfig = byConfig
-	s.invalidateLintProjectCaches()
-	return nil
-}
-
-// reloadConfigAndRelint re-discovers and reloads the rslint JSON config, then
-// re-lints all open documents. The JSON config remains a live fallback for
-// files that have no JS/TS config ancestor, so it must stay current even while
-// one or more JS/TS configs are active.
-func (s *Server) reloadConfigAndRelint() {
-	log.Printf("Reloading rslint config...")
-
-	configPath, found := findRslintConfig(s.fs, s.cwd)
-	if !found {
-		log.Printf("rslint config file no longer exists, clearing config")
-		emptyConfig := config.RslintConfig{}
-		ownerIndex, fileConfigResolver, err := prepareJSONConfigEvaluation(
-			emptyConfig,
-			s.cwd,
-			s.fs,
-		)
-		if err != nil {
-			log.Printf("Error clearing rslint config: %v", err)
-			return
-		}
-		s.jsonConfig = emptyConfig
-		s.jsonConfigOwnerIndex = ownerIndex
-		s.jsonFileConfigResolver = fileConfigResolver
-		s.rslintConfigPath = ""
-		s.tsConfigPaths = nil
-		s.invalidateLintProjectCaches()
-	} else {
-		previousPath := s.rslintConfigPath
-		s.rslintConfigPath = configPath
-		if err := s.reloadConfig(); err != nil {
-			s.rslintConfigPath = previousPath
-			log.Printf("Error reloading rslint config: %v", err)
-			return
-		}
-	}
-
-	for uri := range s.documents {
-		s.pushDiagnostics(uri)
-	}
-}
-
-// lintDebounceDelay is how long to wait after the last keystroke before
-// running the linter. This avoids linting on every keystroke against
-// incomplete/broken syntax that can cause panics or waste CPU.
-const lintDebounceDelay = 200 * time.Millisecond
-
-// scheduleLint marks a URI for deferred linting and resets the debounce timer.
-// When the timer fires it signals debounceCh, which is consumed by the main
-// dispatch loop. Must be called from the main dispatch loop goroutine.
-func (s *Server) scheduleLint(uri lsproto.DocumentUri) {
-	s.pendingLintURIs[uri] = struct{}{}
-	if s.lintTimer != nil {
-		s.lintTimer.Stop()
-	}
-	s.lintTimer = time.AfterFunc(lintDebounceDelay, func() {
-		select {
-		case s.debounceCh <- struct{}{}:
-		default:
-			// Already pending — no need to queue another signal
-		}
-	})
-}
-
-func (s *Server) handleDidOpen(ctx context.Context, params *lsproto.DidOpenTextDocumentParams) error {
-	log.Printf("Handling didOpen: %s", params.TextDocument.Uri)
-
-	uri := params.TextDocument.Uri
-	content := params.TextDocument.Text
-
-	s.documents[uri] = content
-	if s.lintPrograms != nil {
-		existedOnDisk := s.fs != nil && s.fs.FileExists(uriToPath(uri))
-		s.lintPrograms.DidOpen(uri, content, existedOnDisk)
-	}
-
-	// Notify session about the opened file so it creates the overlay.
-	if s.session != nil {
-		s.session.DidOpenFile(ctx, uri, params.TextDocument.Version, content, params.TextDocument.LanguageId)
-		s.pushDiagnostics(uri)
-	}
-	return nil
-}
-
-func (s *Server) handleDidChange(ctx context.Context, params *lsproto.DidChangeTextDocumentParams) error {
-	log.Printf("Handling didChange: %s (version %d)", params.TextDocument.Uri, params.TextDocument.Version)
-
-	uri := params.TextDocument.Uri
-
-	content, err := applyDocumentChanges(s.documents[uri], params.ContentChanges)
-	if err != nil {
-		// didChange is a notification, so the server cannot return an error to
-		// request a retry. Keep both document mirrors on their previous content
-		// instead of partially applying malformed input or sending an invalid
-		// JSON-RPC response with a null request ID.
-		log.Printf("Ignoring invalid didChange for %s: %v", uri, err)
-		return nil
-	}
-	s.documents[uri] = content
-	if s.lintPrograms != nil {
-		s.lintPrograms.DidChange(uri, s.documents[uri])
-	}
-
-	// A content version change supersedes plugin work immediately, not when the
-	// debounce timer eventually starts the next lint. Otherwise an older worker
-	// result can be published against the new buffer while the user is typing.
-	s.docGeneration[uri]++
-	s.cancelInflightPluginDispatch(uri)
-	delete(s.diagnostics, uri)
-
-	// Notify session immediately so tsgo's overlay stays up-to-date for
-	// other LSP features (completions, hover, etc.).  Lint is deferred
-	// via scheduleLint to avoid running the linter on every keystroke.
-	if s.session != nil {
-		s.session.DidChangeFile(ctx, uri, params.TextDocument.Version, params.ContentChanges)
-		s.scheduleLint(uri)
-	}
-	return nil
-}
-
-func (s *Server) handleDidSave(ctx context.Context, params *lsproto.DidSaveTextDocumentParams) error {
-	log.Printf("Handling didSave: %s", params.TextDocument.Uri)
-	uri := params.TextDocument.Uri
-
-	// didChange is authoritative for the current content of an open document.
-	// didSave may include the text that reached disk, but carries no document
-	// version, so a save for an older buffer can arrive after a newer didChange.
-	// Never replace the versioned document mirror with this unversioned snapshot.
-	currentContent, open := s.documents[uri]
-	forwardSave := shouldForwardDidSave(currentContent, open, params.Text)
-	if !forwardSave {
-		log.Printf("Ignoring stale didSave for open document %s", uri)
-	}
-
-	// Clear pending debounce lint for this URI — pushDiagnostics below
-	// will lint it immediately, so the debounce would be redundant.
-	delete(s.pendingLintURIs, uri)
-
-	if s.lintPrograms != nil && forwardSave {
-		s.lintPrograms.DidSave(uri, open)
-	}
-
-	// Notify session about the save event.
-	if s.session != nil {
-		if forwardSave {
-			s.session.DidSaveFile(ctx, uri)
-		}
-		s.pushDiagnostics(uri)
-	}
-	return nil
-}
-
-// shouldForwardDidSave suppresses only saves that are known to describe an
-// older version of an open document. Saves without text and saves for documents
-// not tracked as open are forwarded for LSP client compatibility and so tsgo
-// can observe out-of-band disk changes.
-func shouldForwardDidSave(currentContent string, open bool, savedText *string) bool {
-	return savedText == nil || !open || currentContent == *savedText
-}
-
-func (s *Server) handleDidClose(ctx context.Context, params *lsproto.DidCloseTextDocumentParams) error {
-	log.Printf("Handling didClose: %s", params.TextDocument.Uri)
-	uri := params.TextDocument.Uri
-	delete(s.documents, uri)
-	delete(s.diagnostics, uri)
-	delete(s.pendingLintURIs, uri)
-	// Bump (do NOT delete) the generation on close so any in-flight plugin
-	// result for this URI is stale. Keeping the counter monotonic — rather
-	// than resetting it to 0 on a later reopen — prevents a generation collision
-	// where a pre-close worker result could match a freshly reopened document.
-	s.docGeneration[uri]++
-	// Cancel an in-flight plugin dispatch for the closed doc so its Node worker
-	// stops instead of running to completion — no superseding keystroke will.
-	s.cancelInflightPluginDispatch(uri)
-	if s.lintPrograms != nil {
-		s.lintPrograms.DidClose(uri)
-	}
-
-	if s.session != nil {
-		// Push empty diagnostics to clear the client's display before closing
-		if err := s.PublishDiagnostics(ctx, &lsproto.PublishDiagnosticsParams{
-			Uri:         uri,
-			Diagnostics: []*lsproto.Diagnostic{},
-		}); err != nil {
-			log.Printf("Error clearing diagnostics on close: %v", err)
-		}
-		s.session.DidCloseFile(ctx, uri)
-	}
-	return nil
-}
-
-func (s *Server) handleCodeAction(ctx context.Context, params *lsproto.CodeActionParams) (lsproto.CodeActionResponse, error) {
-	log.Printf("Handling codeAction: %+v,%+v", params, ctx)
-	uri := params.TextDocument.Uri
-
-	// Handle source.fixAll requests (triggered by editor.codeActionsOnSave)
-	if isFixAllRequest(params.Context) {
-		return s.handleFixAllCodeAction(ctx, uri)
-	}
-
-	// Get stored diagnostics for this document
-	ruleDiagnostics, exists := s.diagnostics[uri]
-	if !exists {
-		return lsproto.CodeActionResponse{
-			CommandOrCodeActionArray: &[]lsproto.CommandOrCodeAction{},
-		}, nil
-	}
-
-	var codeActions []lsproto.CommandOrCodeAction
-
-	// Find diagnostics that overlap with the requested range
-	for _, ruleDiag := range ruleDiagnostics {
-		// Check if diagnostic range overlaps with requested range
-		diagStartLine, diagStartChar := scanner.GetECMALineAndUTF16CharacterOfPosition(ruleDiag.SourceFile, ruleDiag.Range.Pos())
-		diagEndLine, diagEndChar := scanner.GetECMALineAndUTF16CharacterOfPosition(ruleDiag.SourceFile, ruleDiag.Range.End())
-
-		diagRange := lsproto.Range{
-			Start: lsproto.Position{Line: uint32(diagStartLine), Character: uint32(diagStartChar)},
-			End:   lsproto.Position{Line: uint32(diagEndLine), Character: uint32(diagEndChar)},
-		}
-
-		if rangesOverlap(diagRange, params.Range) {
-			// Add code action for fixes
-			codeAction := createCodeActionFromRuleDiagnostic(ruleDiag, uri)
-			if codeAction != nil {
-				codeActions = append(codeActions, lsproto.CommandOrCodeAction{
-					Command:    nil,
-					CodeAction: codeAction,
-				})
-			}
-			// add extract disable rule actions
-			disableActions := createDisableRuleActions(ruleDiag, uri)
-			codeActions = append(codeActions, disableActions...)
-
-			// Add code actions for suggestions
-			if ruleDiag.Suggestions != nil {
-				for _, suggestion := range *ruleDiag.Suggestions {
-					suggestionAction := createCodeActionFromSuggestion(ruleDiag, suggestion, uri)
-					if suggestionAction != nil {
-						codeActions = append(codeActions, lsproto.CommandOrCodeAction{
-							Command:    nil,
-							CodeAction: suggestionAction,
-						})
-					}
-				}
-			}
-		}
-	}
-
-	return lsproto.CodeActionResponse{
-		CommandOrCodeActionArray: &codeActions,
-	}, nil
-}
-
-// isFixAllRequest returns true if the code action context requests source.fixAll actions.
-func isFixAllRequest(ctx *lsproto.CodeActionContext) bool {
-	if ctx == nil || ctx.Only == nil {
-		return false
-	}
-	for _, kind := range *ctx.Only {
-		if kind == lsproto.CodeActionKindSourceFixAll || kind == codeActionKindSourceFixAllRslint {
-			return true
-		}
-	}
-	return false
-}
-
-// maxFixPasses is the maximum number of lint-fix cycles to prevent infinite loops
-// when two rules produce fixes that undo each other.
-const maxFixPasses = 10
-
-// handleFixAllCodeAction computes all auto-fixes for the given URI using
-// multi-pass fixing: each pass lints and applies fixes to an isolated overlay,
-// repeating until no more fixes are found or maxFixPasses is reached.
-// This handles cascading fixes (e.g. no-wrapper-object-types fix triggers no-inferrable-types).
-// It does NOT push diagnostics or update s.diagnostics — that is left to the
-// subsequent didSave handler in the normal save flow.
-func (s *Server) handleFixAllCodeAction(ctx context.Context, uri lsproto.DocumentUri) (lsproto.CodeActionResponse, error) {
-	empty := lsproto.CodeActionResponse{CommandOrCodeActionArray: &[]lsproto.CommandOrCodeAction{}}
-
-	// Clear pending debounce for this URI — we are about to lint it fresh,
-	// so any scheduled debounce lint for the same content is redundant.
-	delete(s.pendingLintURIs, uri)
-
-	if s.session == nil {
-		return empty, nil
-	}
-	if !isLintableScriptFile(uri) {
-		return empty, nil
-	}
-	snapshot := s.documentLintSnapshot(uri)
-	if snapshot.unavailable {
-		return empty, nil
-	}
-	originalContent := s.documents[uri]
-
-	currentContent := s.computeFixAllContent(ctx, uri, originalContent, snapshot)
-
-	if currentContent == originalContent {
-		return empty, nil
-	}
-
-	// Produce a single TextEdit that replaces the entire document content.
-	// Individual per-fix TextEdits can't be composed across passes (offsets shift),
-	// so we replace the whole document with the final result.
-	lastLine, lastChar := computeEndPosition(originalContent)
-
-	codeAction := &lsproto.CodeAction{
-		Title: "Fix all rslint auto-fixable problems",
-		Kind:  ptrTo(codeActionKindSourceFixAllRslint),
-		Edit: &lsproto.WorkspaceEdit{
-			Changes: &map[lsproto.DocumentUri][]*lsproto.TextEdit{
-				uri: {
-					{
-						Range: lsproto.Range{
-							Start: lsproto.Position{Line: 0, Character: 0},
-							End:   lsproto.Position{Line: uint32(lastLine), Character: uint32(lastChar)},
-						},
-						NewText: currentContent,
-					},
-				},
-			},
-		},
-	}
-
-	return lsproto.CodeActionResponse{
-		CommandOrCodeActionArray: &[]lsproto.CommandOrCodeAction{
-			{CodeAction: codeAction},
-		},
-	}, nil
-}
-
-// computeFixAllContent runs the multi-pass lint→fix loop and returns the final
-// fixed content (== originalContent when nothing changed). Each pass folds
-// eslint-plugin fixes into the native fixes so source.fixAll applies both, on
-// the SAME content (so their byte offsets align with ApplyRuleFixes's input).
-// The per-pass native lint goes through s.fixAllNativeLint, which tests override
-// to drive the fold loop without a real TS session.
-func (s *Server) computeFixAllContent(
-	ctx context.Context,
-	uri lsproto.DocumentUri,
-	originalContent string,
-	snapshot documentLintSnapshot,
-) string {
-	nativeLint := s.fixAllNativeLint
-	if nativeLint == nil {
-		nativeLint = s.defaultFixAllNativeLint
-	}
-
-	// Bound the eslint-plugin reverse requests across the WHOLE fixAll, not per
-	// pass: source.fixAll runs inline on the dispatch loop, so a wedged or
-	// mid-rebuild client that never answers rslint/pluginLint must not
-	// freeze editor interaction — nor multiply the stall by maxFixPasses. Only
-	// the plugin pass gets this deadline; the native pass keeps the original ctx
-	// (it is in-process and does not depend on a client reply). Once the budget
-	// expires lintPluginRulesSync returns nil and the remaining passes fold
-	// native-only fixes.
-	pluginTimeout := s.pluginReverseTimeout
-	if pluginTimeout <= 0 {
-		pluginTimeout = defaultPluginReverseTimeout
-	}
-	pluginCtx, cancelPlugin := context.WithTimeout(ctx, pluginTimeout)
-	defer cancelPlugin()
-
-	currentContent := originalContent
-	for pass := range maxFixPasses {
-		lintResult, err := nativeLint(ctx, uri, pass, currentContent, snapshot)
-		if err != nil {
-			log.Printf("Error running lint for fixAll pass %d: %v", pass, err)
-			break
-		}
-		if lintResult.HasSyntaxErrors {
-			break
-		}
-		ruleDiags := lintResult.Diagnostics
-
-		// Fold in eslint-plugin fixes so source.fixAll applies plugin rule fixes
-		// too, not just native. The plugin pass lints the SAME currentContent, so
-		// its fix byte offsets align with ApplyRuleFixes's input; suggestionsMode
-		// is "off" because fixAll applies only autofixes.
-		// Skip the plugin pass once the budget is spent: lintPluginRulesSync on an
-		// already-expired pluginCtx would still enqueue a (wasted) reverse request
-		// to the client before returning nil.
-		if pluginCtx.Err() == nil {
-			if pluginDiags := s.lintPluginRulesSyncWithSnapshot(
-				pluginCtx,
-				uri,
-				currentContent,
-				true,
-				linter.SuggestionsModeOff,
-				snapshot,
-			); len(pluginDiags) > 0 {
-				ruleDiags = append(ruleDiags, pluginDiags...)
-			}
-		}
-
-		fixedContent, _, wasFixed := linter.ApplyRuleFixes(currentContent, ruleDiags)
-		if !wasFixed {
-			break
-		}
-		currentContent = fixedContent
-		if currentContent == originalContent {
-			break // cycle detected — fixes reverted to original content
-		}
-	}
-	return currentContent
-}
-
-// defaultFixAllNativeLint builds each pass from an isolated editor overlay.
-// The pass number is intentionally unused: speculative content never enters
-// the real TypeScript Session, regardless of how many fix cycles run.
-func (s *Server) defaultFixAllNativeLint(
-	ctx context.Context,
-	uri lsproto.DocumentUri,
-	_ int,
-	content string,
-	snapshot documentLintSnapshot,
-) (lintPassResult, error) {
-	return s.runConfiguredLintForContentWithSnapshot(uri, ctx, content, snapshot)
-}
-
-// computeEndPosition returns the line and UTF-16 character offset of the end
-// of a text string, suitable for constructing an LSP Range that covers the
-// entire document. Uses core.UTF16Len for correct UTF-16 code unit counting.
-func computeEndPosition(text string) (int, int) {
-	line := 0
-	lastLineStart := 0
-	for i := range len(text) {
-		if text[i] == '\n' {
-			line++
-			lastLineStart = i + 1
-		}
-	}
-	return line, int(core.UTF16Len(text[lastLineStart:]))
-}
-
-func convertRuleDiagnosticToLSP(ruleDiag rule.RuleDiagnostic) *lsproto.Diagnostic {
-	diagnosticStart := ruleDiag.Range.Pos()
-	diagnosticEnd := ruleDiag.Range.End()
-	startLine, startColumn := scanner.GetECMALineAndUTF16CharacterOfPosition(ruleDiag.SourceFile, diagnosticStart)
-	endLine, endColumn := scanner.GetECMALineAndUTF16CharacterOfPosition(ruleDiag.SourceFile, diagnosticEnd)
-
-	return &lsproto.Diagnostic{
-		Range: lsproto.Range{
-			Start: lsproto.Position{
-				Line:      uint32(startLine),
-				Character: uint32(startColumn),
-			},
-			End: lsproto.Position{
-				Line:      uint32(endLine),
-				Character: uint32(endColumn),
-			},
-		},
-		Severity: ptrTo(lsproto.DiagnosticSeverity(ruleDiag.Severity.Int())),
-		Source:   ptrTo("rslint"),
-		Message:  lsproto.StringOrMarkupContent{String: ptrTo(fmt.Sprintf("[%s] %s", ruleDiag.RuleName, ruleDiag.Message.Description))},
-	}
-}
-
-func isLintableScriptFile(uri lsproto.DocumentUri) bool {
-	return config.IsSupportedLintFile(uriToPath(uri))
-}
-
-func uriToPath(uri lsproto.DocumentUri) string {
-	// Convert file:// URI to file path using net/url for proper percent-decoding.
-	// Handles spaces (%20), CJK characters, and other encoded chars in paths.
-	// file:///home/user       → /home/user  (Unix)
-	// file:///C:/Users        → C:/Users    (Windows — strip the leading slash)
-	// file:///path%20name/f   → /path name/f
-	uriStr := string(uri)
-	if uriStr == "" {
-		return ""
-	}
-	u, err := url.ParseRequestURI(uriStr)
-	if err != nil {
-		return uriStr // fallback: return as-is for non-URI strings
-	}
-	p := u.Path
-	if u.Host != "" {
-		return "//" + u.Host + p
-	}
-	// Windows drive letter: /C:/... → C:/...
-	if len(p) >= 3 && p[0] == '/' && unicode.IsLetter(rune(p[1])) && p[2] == ':' {
-		return p[1:]
-	}
-	return p
-}
-
-func findRslintConfig(fs vfs.FS, workingDir string) (string, bool) {
-	defaultConfigs := []string{"rslint.json", "rslint.jsonc"}
-
-	// Strategy 1: Try in the working directory
-	for _, configName := range defaultConfigs {
-		configPath := filepath.Join(workingDir, configName)
-		if fs.FileExists(configPath) {
-			return configPath, true
-		}
-	}
-	return "", false
-}
-
-// LintResponse represents a lint response from Go to JS
-type LintResponse struct {
-	Diagnostics []lsproto.Diagnostic `json:"diagnostics"`
-	ErrorCount  int                  `json:"errorCount"`
-	FileCount   int                  `json:"fileCount"`
-	RuleCount   int                  `json:"ruleCount"`
-}
-
-func runLintWithSession(uri lsproto.DocumentUri, session *project.Session, ctx context.Context, rslintConfig config.RslintConfig, cwd string, enforcePlugins bool, tsConfigPaths []string, fs vfs.FS) ([]rule.RuleDiagnostic, error) {
-	target := lspConfigTarget(uriToPath(uri), cwd, fs)
-	snapshot := resolveDocumentLintSnapshotConfig(documentLintSnapshot{
-		target: target,
-		config: rslintConfig,
-		pathSpaces: config.NewPathSpaceSnapshot(
-			map[string]config.RslintConfig{target.ConfigDirectory: rslintConfig},
-			fs,
-		),
-		ruleCatalog:           rules.All(),
-		typeScriptConfigPaths: tsConfigPaths,
-		usesJavaScriptConfig:  enforcePlugins,
-	}, fs)
-	result, err := runLintWithProgramLoader(
-		uri,
-		session,
-		ctx,
-		snapshot,
-		cwd,
-		fs,
-		lintProjectLoaders{},
-		nil,
-	)
-	return result.Diagnostics, err
-}
-
-// runLintWithProgramLoader resolves one document against two distinct
-// directories: configCwd is the config's own path space, which a nested JS
-// config moves to its own directory, while processCwd is the server's working
-// directory that rules see through RuleContext.ProcessCurrentDirectory.
-func runLintWithProgramLoader(
-	uri lsproto.DocumentUri,
-	session *project.Session,
-	ctx context.Context,
-	snapshot documentLintSnapshot,
-	processCwd string,
-	fs vfs.FS,
-	loaders lintProjectLoaders,
-	sessionRoots *lintSessionProjectRootCache,
-) (lintPassResult, error) {
-	if !snapshot.configResolved {
-		snapshot = resolveDocumentLintSnapshotConfig(snapshot, fs)
-	}
-	target := snapshot.target
-	if isDefaultExcludedLintPath(target.Path, processCwd, fs) {
-		return lintPassResult{Diagnostics: []rule.RuleDiagnostic{}}, nil
-	}
-
-	// Files excluded by the config's `ignores` patterns produce no diagnostics,
-	// matching CLI behavior. Return early before spinning up the language service.
-	if snapshot.resolvedConfig.GloballyIgnored {
-		return lintPassResult{Diagnostics: []rule.RuleDiagnostic{}}, nil
-	}
-
-	program, sourceFile, hasTypeInfo, err := selectLintProgram(
-		uri,
-		target,
-		session,
-		ctx,
-		snapshot.typeScriptConfigPaths,
-		fs,
-		loaders,
-		sessionRoots,
-	)
-	if err != nil {
-		return lintPassResult{}, err
-	}
-	return lintSingleFile(
-		program,
-		sourceFile,
-		target,
-		processCwd,
-		hasTypeInfo,
-		snapshot.resolvedConfig.EnabledRules,
-		rule.EditDemandAll,
-		ctx,
-	), nil
-}
-
-// rulesSkippedInEditors names the rules the language server never runs. A rule
-// belongs here when what it checks is a property of the file's bytes rather
-// than of its text: an editor's document holds text that has already been
-// decoded, so such a property is neither visible in the document nor reachable
-// by a text edit.
-var rulesSkippedInEditors = map[string]bool{
-	// unicode-bom checks for a leading byte order mark. An editor decodes the
-	// mark into the document's encoding — VS Code shows it in the status bar as
-	// "UTF-8 with BOM" — so the document text never carries it and no text edit
-	// adds or removes it. The file on disk is the only remaining witness, and an
-	// unsaved buffer may already disagree with it. `rslint --fix`, which rewrites
-	// the file itself, is where the rule applies.
-	"unicode-bom": true,
-}
-
-// rulesServedToEditors drops the rules the language server never runs. The
-// input is a cached slice shared across files, so filtering builds a new one
-// and an unaffected configuration keeps the original.
-func rulesServedToEditors(rules []rule.ConfiguredRule) []rule.ConfiguredRule {
-	skipped := func(r rule.ConfiguredRule) bool { return rulesSkippedInEditors[r.Name] }
-	if !slices.ContainsFunc(rules, skipped) {
-		return rules
-	}
-	served := make([]rule.ConfiguredRule, 0, len(rules))
-	for _, configured := range rules {
-		if !skipped(configured) {
-			served = append(served, configured)
-		}
-	}
-	return served
-}
-
-func lintSingleFile(
-	program *compiler.Program,
-	sourceFile *ast.SourceFile,
-	target target.File,
-	processCwd string,
-	hasTypeInfo bool,
-	enabledRules []rule.ConfiguredRule,
-	editDemand rule.EditDemand,
-	ctx context.Context,
-) lintPassResult {
-	if sourceFile == nil {
-		return lintPassResult{Diagnostics: []rule.RuleDiagnostic{}}
-	}
-	sourceProgram := lintprogram.NewFromCompiler(program)
-	if syntacticDiagnostics := sourceProgram.SyntacticDiagnostics(ctx, sourceFile); len(syntacticDiagnostics) > 0 {
-		diagnostics := make([]rule.RuleDiagnostic, 0, len(syntacticDiagnostics))
-		for _, diagnostic := range syntacticDiagnostics {
-			diagnostics = append(diagnostics, rule.RuleDiagnostic{
-				RuleName:     fmt.Sprintf("TypeScript(TS%d)", diagnostic.Code()),
-				SourceFile:   sourceFile,
-				FilePath:     sourceFile.FileName(),
-				Range:        diagnostic.Loc(),
-				Message:      rule.RuleMessage{Description: diagnostic.String()},
-				Severity:     rule.SeverityError,
-				Origin:       rule.DiagnosticOriginTypeScript,
-				PreFormatted: true,
-			})
-		}
-		return lintPassResult{Diagnostics: diagnostics, HasSyntaxErrors: true}
-	}
-
-	// Collect diagnostics
-	var diagnostics []rule.RuleDiagnostic
-	var diagnosticsLock sync.Mutex
-
-	// Create collector function
-	diagnosticCollector := func(d rule.RuleDiagnostic) {
-		diagnosticsLock.Lock()
-		defer diagnosticsLock.Unlock()
-		diagnostics = append(diagnostics, d)
-	}
-
-	linter.LintSingleFile(linter.LintSingleFileOptions{
-		Program:     sourceProgram,
-		File:        sourceFile.FileName(),
-		Cwd:         processCwd,
-		HasTypeInfo: hasTypeInfo,
-		GetRulesForFile: func(*ast.SourceFile) []rule.ConfiguredRule {
-			return rulesServedToEditors(enabledRules)
-		},
-		Consumer: rule.DiagnosticConsumer{
-			Demand: editDemand,
-			Report: diagnosticCollector,
-		},
-	})
-
-	if diagnostics == nil {
-		diagnostics = []rule.RuleDiagnostic{}
-	}
-	return lintPassResult{Diagnostics: diagnostics}
-}
-
-func selectLintProgram(
-	uri lsproto.DocumentUri,
-	target target.File,
-	session *project.Session,
-	ctx context.Context,
-	tsConfigPaths []string,
-	fs vfs.FS,
-	fallbackLoaders lintProjectLoaders,
-	sessionRoots *lintSessionProjectRootCache,
-) (*compiler.Program, *ast.SourceFile, bool, error) {
-	// Flush pending document changes and collect every already-loaded project
-	// containing the file. The default language service remains the
-	// non-project-backed fallback when none of the config's declared projects
-	// contains the file.
-	_, languageService, loadedProjects, err := session.GetLanguageServiceAndProjectsForFile(ctx, uri)
-	if err != nil {
-		return nil, nil, false, fmt.Errorf("failed to get language service: %w", err)
-	}
-	program := languageService.GetProgram()
-
-	// Type information follows parserOptions.project declaration order, not the
-	// TypeScript session's default-project heuristic. Direct config roots outrank
-	// files present only through imports. Custom config names that the main
-	// project service has not loaded are supplied by rslint-owned
-	// session-external ts-go Programs.
-	type loadedLintProject struct {
-		program     *compiler.Program
-		commandLine *tsoptions.ParsedCommandLine
-	}
-	loadedByConfig := make(map[tspath.Path]loadedLintProject, len(loadedProjects))
-	for _, candidate := range loadedProjects {
-		if candidate == nil || candidate.GetProgram() == nil {
-			continue
-		}
-		candidateProgram := candidate.GetProgram()
-		configPath := string(candidate.Id())
-		if configPath == "" {
-			continue
-		}
-		commandLine := candidateProgram.CommandLine()
-		if sessionProject, ok := candidate.(*project.Project); ok && sessionProject.CommandLine != nil {
-			// The Program command line may include automatic type-acquisition
-			// roots. Project.CommandLine is the authored tsconfig root set.
-			commandLine = sessionProject.CommandLine
-		}
-		loadedByConfig[lintProgramLexicalPathID(configPath, fs)] = loadedLintProject{
-			program:     candidateProgram,
-			commandLine: commandLine,
-		}
-	}
-	loaders := lintProjectLoaders{
-		metadata: func(tsConfigPath string) (*lintProjectMetadata, bool, error) {
-			if loadedProject, ok := loadedByConfig[lintProgramLexicalPathID(tsConfigPath, fs)]; ok {
-				metadata := sessionRoots.metadata(tsConfigPath, loadedProject.commandLine, fs)
-				return metadata, metadata != nil, nil
-			}
-			if fallbackLoaders.metadata == nil {
-				return nil, false, nil
-			}
-			return fallbackLoaders.metadata(tsConfigPath)
-		},
-		program: func(tsConfigPath string) (*compiler.Program, *ast.SourceFile, error) {
-			if loadedProject, ok := loadedByConfig[lintProgramLexicalPathID(tsConfigPath, fs)]; ok {
-				return loadedProject.program, sourceFileForTarget(loadedProject.program, target, fs), nil
-			}
-			if fallbackLoaders.program == nil {
-				return nil, nil, nil
-			}
-			return fallbackLoaders.program(tsConfigPath)
-		},
-	}
-	selected, found, err := selectConfiguredLintProject(tsConfigPaths, target, loaders)
-	if err != nil {
-		return nil, nil, false, err
-	}
-	if found {
-		return selected.program, selected.sourceFile, true, nil
-	}
-	return program, sourceFileForTarget(program, target, fs), false, nil
-}
-
-func sourceFileForPath(program *compiler.Program, filename string, fs vfs.FS) *ast.SourceFile {
-	return utils.NewProgramSourceLookup(program, fs).SourceFileForPath(filename)
-}
-
-func sourceFileForTarget(
-	program *compiler.Program,
-	target target.File,
-	fs vfs.FS,
-) *ast.SourceFile {
-	return utils.NewProgramSourceLookup(program, fs).
-		SourceFileForTarget(target.Path, target.CanonicalPath)
-}
-
-func (s *Server) currentEditorOverlayFSForTarget(
-	uri lsproto.DocumentUri,
-	target target.File,
-) vfs.FS {
-	content, open := s.documents[uri]
-	files, _ := s.currentEditorOverlayFilesForFrozenTarget(
-		uri,
-		target,
-		content,
-		open,
-	)
-	return newFrozenLintTargetOverlayFS(s.fs, files, target)
-}
-
-func (s *Server) currentEditorOverlayFSForTargetWithConflicts(
-	uri lsproto.DocumentUri,
-	target target.File,
-) (vfs.FS, bool) {
-	content, open := s.documents[uri]
-	files, aliasesConflict := s.currentEditorOverlayFilesForFrozenTarget(
-		uri,
-		target,
-		content,
-		open,
-	)
-	return newFrozenLintTargetOverlayFS(s.fs, files, target), aliasesConflict
-}
-
-func (s *Server) currentEditorOverlayFilesForTarget(
-	uri lsproto.DocumentUri,
-	target target.File,
-	content string,
-) map[string]string {
-	files, _ := s.currentEditorOverlayFilesForFrozenTarget(
-		uri,
-		target,
-		content,
-		true,
-	)
-	return files
-}
-
-// currentEditorOverlayFilesForFrozenTarget resolves every other open document
-// normally, but inserts the selected document only under the identity frozen
-// by documentLintSnapshot. Re-resolving the selected URI here could mix two
-// symlink generations between config/project selection and Program creation.
-func (s *Server) currentEditorOverlayFilesForFrozenTarget(
-	uri lsproto.DocumentUri,
-	target target.File,
-	targetContent string,
-	includeTarget bool,
-) (map[string]string, bool) {
-	files := make(map[string]string, len(s.documents)*2)
-	contentByPhysicalPath := make(map[string]string, len(s.documents))
-	aliasesConflict := false
-	add := func(filePath string, content string) {
-		physicalPath := s.addEditorOverlayFile(files, filePath, content)
-		if previous, exists := contentByPhysicalPath[physicalPath]; exists &&
-			previous != content {
-			aliasesConflict = true
-		}
-		contentByPhysicalPath[physicalPath] = content
-	}
-	for documentURI, documentContent := range s.documents {
-		if documentURI == uri {
-			continue
-		}
-		add(uriToPath(documentURI), documentContent)
-	}
-	if includeTarget {
-		physicalPath := frozenLintTargetPhysicalPathID(target, s.fs)
-		if previous, exists := contentByPhysicalPath[physicalPath]; exists &&
-			previous != targetContent {
-			aliasesConflict = true
-		}
-		contentByPhysicalPath[physicalPath] = targetContent
-		addEditorOverlayTarget(files, target, targetContent)
-	}
-	return files, aliasesConflict
-}
-
-func frozenLintTargetPhysicalPathID(
-	target target.File,
-	fs vfs.FS,
-) string {
-	filePath := target.CanonicalPath
-	if filePath == "" {
-		filePath = target.Path
-	}
-	caseSensitive := true
-	if fs != nil {
-		caseSensitive = fs.UseCaseSensitiveFileNames()
-	}
-	return lspLexicalPathID(filePath, caseSensitive)
-}
-
-type frozenLintTargetOverlayFS struct {
-	vfs.FS
-	lexicalPathID   string
-	canonicalPathID string
-	canonicalPath   string
-}
-
-func newFrozenLintTargetOverlayFS(
-	baseFS vfs.FS,
-	files map[string]string,
-	target target.File,
-) vfs.FS {
-	caseSensitive := true
-	if baseFS != nil {
-		caseSensitive = baseFS.UseCaseSensitiveFileNames()
-	}
-	canonicalPath := target.CanonicalPath
-	if canonicalPath == "" {
-		canonicalPath = target.Path
-	}
-	canonicalPath = tspath.NormalizePath(canonicalPath)
-	return &frozenLintTargetOverlayFS{
-		FS:              utils.NewOverlayVFS(baseFS, files),
-		lexicalPathID:   lspLexicalPathID(target.Path, caseSensitive),
-		canonicalPathID: lspLexicalPathID(canonicalPath, caseSensitive),
-		canonicalPath:   canonicalPath,
-	}
-}
-
-func (fs *frozenLintTargetOverlayFS) Realpath(filePath string) string {
-	caseSensitive := fs.UseCaseSensitiveFileNames()
-	pathID := lspLexicalPathID(filePath, caseSensitive)
-	if pathID == fs.lexicalPathID || pathID == fs.canonicalPathID {
-		return fs.canonicalPath
-	}
-	return fs.FS.Realpath(filePath)
-}
-
-func addEditorOverlayTarget(
-	files map[string]string,
-	target target.File,
-	content string,
-) {
-	if target.Path != "" {
-		files[tspath.NormalizePath(target.Path)] = content
-	}
-	if target.CanonicalPath != "" {
-		files[tspath.NormalizePath(target.CanonicalPath)] = content
-	}
-}
-
-func (s *Server) addEditorOverlayFile(
-	files map[string]string,
-	filePath string,
-	content string,
-) string {
-	filePath = tspath.NormalizePath(filePath)
-	files[filePath] = content
-	caseSensitive := true
-	if s.fs != nil {
-		caseSensitive = s.fs.UseCaseSensitiveFileNames()
-	}
-	if s.fs == nil {
-		return string(tspath.ToPath(filePath, "", caseSensitive))
-	}
-	if realPath := s.fs.Realpath(filePath); realPath != "" {
-		realPath = tspath.NormalizePath(realPath)
-		files[realPath] = content
-		return string(tspath.ToPath(realPath, "", caseSensitive))
-	}
-	return string(tspath.ToPath(filePath, "", caseSensitive))
-}
-
-func createStandaloneFallbackProgram(filename string, cwd string, fs vfs.FS) (*compiler.Program, error) {
-	host := utils.CreateCompilerHost(cwd, fs)
-	return utils.CreateProgramFromOptionsLenient(true, &core.CompilerOptions{
-		Target:    core.ScriptTargetESNext,
-		Module:    core.ModuleKindESNext,
-		Jsx:       core.JsxEmitPreserve,
-		AllowJs:   core.TSTrue,
-		NoLib:     core.TSTrue,
-		NoResolve: core.TSTrue,
-	}, []string{filename}, host)
-}
-
-func (s *Server) runConfiguredLint(
-	uri lsproto.DocumentUri,
-	ctx context.Context,
-	snapshot documentLintSnapshot,
-) (lintPassResult, error) {
-	request := newStandaloneLintProjectRequest(
-		snapshot.target,
-		func() vfs.FS { return s.currentEditorOverlayFSForTarget(uri, snapshot.target) },
-	)
-	loaders := request.loaders()
-	finalize := func() {}
-	if s.lintPrograms != nil && s.lintPrograms.Usable() {
-		loadProgram, loadMetadata, requestFinalize := s.lintPrograms.Request(
-			ctx,
-			uri,
-			snapshot.target,
-		)
-		loaders = lintProjectLoaders{
-			program:  loadProgram,
-			metadata: loadMetadata,
-		}
-		finalize = requestFinalize
-	}
-	defer finalize()
-	return runLintWithProgramLoader(
-		uri,
-		s.session,
-		ctx,
-		snapshot,
-		s.cwd,
-		s.fs,
-		loaders,
-		s.lintSessionRoots,
-	)
-}
-
-// runConfiguredLintForContent lints a speculative fix pass against an
-// isolated overlay. It never mutates the TypeScript Session's open document,
-// so cancelling or declining a code action cannot change later LSP results.
-func (s *Server) runConfiguredLintForContent(
-	uri lsproto.DocumentUri,
-	ctx context.Context,
-	content string,
-	rslintConfig config.RslintConfig,
-	cwd string,
-	enforcePlugins bool,
-	tsConfigPaths []string,
-) (lintPassResult, error) {
-	ruleCatalog := rules.All()
-	if enforcePlugins {
-		ruleCatalog = s.currentRuleCatalog()
-	}
-	target := lspConfigTarget(uriToPath(uri), cwd, s.fs)
-	snapshot := resolveDocumentLintSnapshotConfig(documentLintSnapshot{
-		target: target,
-		config: rslintConfig,
-		pathSpaces: config.NewPathSpaceSnapshot(
-			map[string]config.RslintConfig{target.ConfigDirectory: rslintConfig},
-			s.fs,
-		),
-		ruleCatalog:           ruleCatalog,
-		typeScriptConfigPaths: tsConfigPaths,
-		usesJavaScriptConfig:  enforcePlugins,
-	}, s.fs)
-	return s.runConfiguredLintForContentWithSnapshot(uri, ctx, content, snapshot)
-}
-
-func (s *Server) runConfiguredLintForContentWithSnapshot(
-	uri lsproto.DocumentUri,
-	ctx context.Context,
-	content string,
-	snapshot documentLintSnapshot,
-) (lintPassResult, error) {
-	if !snapshot.configResolved {
-		snapshot = resolveDocumentLintSnapshotConfig(snapshot, s.fs)
-	}
-	target := snapshot.target
-	cwd := target.ConfigDirectory
-	if isDefaultExcludedLintPath(target.Path, s.cwd, s.fs) {
-		return lintPassResult{Diagnostics: []rule.RuleDiagnostic{}}, nil
-	}
-	if snapshot.resolvedConfig.GloballyIgnored {
-		return lintPassResult{Diagnostics: []rule.RuleDiagnostic{}}, nil
-	}
-
-	// Apply the speculative target last so it wins over every open URI alias
-	// that resolves to the same physical file, without resolving the frozen
-	// target through the filesystem again.
-	files := s.currentEditorOverlayFilesForTarget(uri, target, content)
-	overlayFS := newFrozenLintTargetOverlayFS(s.fs, files, target)
-
-	request := newStandaloneLintProjectRequestWithFS(target, overlayFS)
-	selected, found, err := selectConfiguredLintProject(
-		snapshot.typeScriptConfigPaths,
-		target,
-		request.loaders(),
-	)
-	if err != nil {
-		return lintPassResult{}, err
-	}
-	if found {
-		return lintSingleFile(
-			selected.program,
-			selected.sourceFile,
-			target,
-			s.cwd,
-			true,
-			snapshot.resolvedConfig.EnabledRules,
-			rule.EditDemandAutofix,
-			ctx,
-		), nil
-	}
-
-	program, err := createStandaloneFallbackProgram(target.Path, cwd, overlayFS)
-	if err != nil {
-		return lintPassResult{}, fmt.Errorf("create fallback lint program: %w", err)
-	}
-	return lintSingleFile(
-		program,
-		sourceFileForTarget(program, target, overlayFS),
-		target,
-		s.cwd,
-		false,
-		snapshot.resolvedConfig.EnabledRules,
-		rule.EditDemandAutofix,
-		ctx,
-	), nil
-}
-
-func lspFilesystemPathID(filePath string, fs vfs.FS) string {
-	filePath = tspath.NormalizePath(filePath)
-	caseSensitive := true
-	if fs != nil {
-		caseSensitive = fs.UseCaseSensitiveFileNames()
-		if realPath := fs.Realpath(filePath); realPath != "" {
-			filePath = tspath.NormalizePath(realPath)
-		}
-	}
-	return lspLexicalPathID(filePath, caseSensitive)
-}
-
-func lspTargetIdentity(filePath string, fs vfs.FS) target.File {
-	return target.File{
-		PathIdentity: target.FreezeFileIdentity(filePath, fs),
-	}
-}
-
-func lspConfigTarget(filePath string, configDirectory string, fs vfs.FS) target.File {
-	target := lspTargetIdentity(filePath, fs)
-	target.ConfigDirectory = tspath.NormalizePath(configDirectory)
-	return target
-}
-
-func isDefaultExcludedLintPath(filePath string, cwd string, fs vfs.FS) bool {
-	useCaseSensitive := true
-	if fs != nil {
-		useCaseSensitive = fs.UseCaseSensitiveFileNames()
-	}
-	return config.IsDefaultExcludedPath(filePath, cwd, useCaseSensitive)
-}
-
-// Helper function to check if two ranges overlap
-func rangesOverlap(a, b lsproto.Range) bool {
-	// Ranges overlap if a starts before or at b's end AND b starts before or at a's end
-	aStartsBefore := a.Start.Line < b.End.Line ||
-		(a.Start.Line == b.End.Line && a.Start.Character <= b.End.Character)
-	bStartsBefore := b.Start.Line < a.End.Line ||
-		(b.Start.Line == a.End.Line && b.Start.Character <= a.End.Character)
-
-	return aStartsBefore && bStartsBefore
-}
-
-// Helper function to create a code action from a rule diagnostic
-func createCodeActionFromRuleDiagnostic(ruleDiag rule.RuleDiagnostic, uri lsproto.DocumentUri) *lsproto.CodeAction {
-	fixes := ruleDiag.Fixes()
-	if len(fixes) == 0 {
-		return nil
-	}
-
-	// Convert rule fixes to LSP text edits
-	var textEdits []*lsproto.TextEdit
-	for _, fix := range fixes {
-		textEdits = append(textEdits, ruleFixToTextEdit(ruleDiag.SourceFile, fix))
-	}
-
-	// Create workspace edit
-	workspaceEdit := &lsproto.WorkspaceEdit{
-		Changes: &map[lsproto.DocumentUri][]*lsproto.TextEdit{
-			uri: textEdits,
-		},
-	}
-
-	return &lsproto.CodeAction{
-		Title:       "Fix: " + ruleDiag.Message.Description,
-		Kind:        ptrTo(lsproto.CodeActionKind("quickfix")),
-		Edit:        workspaceEdit,
-		Diagnostics: &[]*lsproto.Diagnostic{convertRuleDiagnosticToLSP(ruleDiag)},
-		IsPreferred: ptrTo(true), // Mark auto-fixes as preferred
-	}
-}
-
-// Helper function to create a code action from a rule suggestion
-func createCodeActionFromSuggestion(ruleDiag rule.RuleDiagnostic, suggestion rule.RuleSuggestion, uri lsproto.DocumentUri) *lsproto.CodeAction {
-	fixes := suggestion.Fixes()
-	if len(fixes) == 0 {
-		return nil
-	}
-
-	// Convert rule fixes to LSP text edits
-	var textEdits []*lsproto.TextEdit
-	for _, fix := range fixes {
-		textEdits = append(textEdits, ruleFixToTextEdit(ruleDiag.SourceFile, fix))
-	}
-
-	// Create workspace edit
-	workspaceEdit := &lsproto.WorkspaceEdit{
-		Changes: &map[lsproto.DocumentUri][]*lsproto.TextEdit{
-			uri: textEdits,
-		},
-	}
-
-	return &lsproto.CodeAction{
-		Title:       "Suggestion: " + suggestion.Message.Description,
-		Kind:        ptrTo(lsproto.CodeActionKind("quickfix")),
-		Edit:        workspaceEdit,
-		Diagnostics: &[]*lsproto.Diagnostic{convertRuleDiagnosticToLSP(ruleDiag)},
-		IsPreferred: ptrTo(false), // Mark suggestions as not preferred
-	}
-}
-
-// Helper function to create disable rule actions for diagnostics without fixes
-func createDisableRuleActions(ruleDiag rule.RuleDiagnostic, uri lsproto.DocumentUri) []lsproto.CommandOrCodeAction {
-	if ruleDiag.Origin == rule.DiagnosticOriginTypeScript {
-		return nil
-	}
-	var actions []lsproto.CommandOrCodeAction
-
-	lspDiagnostic := convertRuleDiagnosticToLSP(ruleDiag)
-
-	// Action 1: Disable rule for this line
-	disableLineAction := createDisableRuleForLineAction(ruleDiag, uri, lspDiagnostic)
-	if disableLineAction != nil {
-		actions = append(actions, lsproto.CommandOrCodeAction{
-			Command:    nil,
-			CodeAction: disableLineAction,
-		})
-	}
-
-	// Action 2: Disable rule for entire file
-	disableFileAction := createDisableRuleForFileAction(ruleDiag, uri, lspDiagnostic)
-	if disableFileAction != nil {
-		actions = append(actions, lsproto.CommandOrCodeAction{
-			Command:    nil,
-			CodeAction: disableFileAction,
-		})
-	}
-
-	return actions
-}
-
-// Helper function to create a "disable rule for this line" action
-func createDisableRuleForLineAction(ruleDiag rule.RuleDiagnostic, uri lsproto.DocumentUri, lspDiagnostic *lsproto.Diagnostic) *lsproto.CodeAction {
-	// Get the line where the diagnostic occurs
-	lineStart := lspDiagnostic.Range.Start.Line
-
-	// Create text edit to add rslint-disable-next-line comment
-	disableComment := fmt.Sprintf("// rslint-disable-next-line %s\n", ruleDiag.RuleName)
-
-	// Find the start of the line to insert the comment
-	lineStartPos := lsproto.Position{Line: lineStart, Character: 0}
-
-	textEdit := &lsproto.TextEdit{
-		Range: lsproto.Range{
-			Start: lineStartPos,
-			End:   lineStartPos,
-		},
-		NewText: disableComment,
-	}
-
-	workspaceEdit := &lsproto.WorkspaceEdit{
-		Changes: &map[lsproto.DocumentUri][]*lsproto.TextEdit{
-			uri: {textEdit},
-		},
-	}
-
-	return &lsproto.CodeAction{
-		Title:       fmt.Sprintf("Disable %s for this line", ruleDiag.RuleName),
-		Kind:        ptrTo(lsproto.CodeActionKind("quickfix")),
-		Edit:        workspaceEdit,
-		Diagnostics: &[]*lsproto.Diagnostic{lspDiagnostic},
-		IsPreferred: ptrTo(false),
-	}
-}
-
-// Helper function to create a "disable rule for entire file" action
-func createDisableRuleForFileAction(ruleDiag rule.RuleDiagnostic, uri lsproto.DocumentUri, lspDiagnostic *lsproto.Diagnostic) *lsproto.CodeAction {
-	// Create text edit to add rslint-disable comment at the top of the file
-	disableComment := fmt.Sprintf("/* rslint-disable %s */\n", ruleDiag.RuleName)
-
-	// Insert at the very beginning of the file
-	fileStartPos := lsproto.Position{Line: 0, Character: 0}
-
-	textEdit := &lsproto.TextEdit{
-		Range: lsproto.Range{
-			Start: fileStartPos,
-			End:   fileStartPos,
-		},
-		NewText: disableComment,
-	}
-
-	workspaceEdit := &lsproto.WorkspaceEdit{
-		Changes: &map[lsproto.DocumentUri][]*lsproto.TextEdit{
-			uri: {textEdit},
-		},
-	}
-
-	return &lsproto.CodeAction{
-		Title:       fmt.Sprintf("Disable %s for entire file", ruleDiag.RuleName),
-		Kind:        ptrTo(lsproto.CodeActionKind("quickfix")),
-		Edit:        workspaceEdit,
-		Diagnostics: &[]*lsproto.Diagnostic{lspDiagnostic},
-		IsPreferred: ptrTo(false),
-	}
-}
-
-type documentConfigSelection struct {
-	entries       config.RslintConfig
-	resolved      config.ResolvedFileConfig
-	pathSpaces    *config.PathSpaceSnapshot
-	ruleCatalog   *rule.Catalog
-	directory     string
-	configKey     string
-	usesJSConfig  bool
-	configMissing bool
-}
-
-func (s *Server) selectDocumentConfig(lintFile target.File) documentConfigSelection {
-	jsRuleCatalog := s.currentRuleCatalog()
-	evaluationFS := s.fs
-	jsOwnerIndex := s.jsConfigOwnerIndex
-	jsFileConfigResolvers := s.jsFileConfigResolvers
-	if !s.configSnapshotIncludesGitignore {
-		// The target identity was frozen before entering this function. Bootstrap
-		// config evaluation now gets its own short-lived filesystem generation so
-		// owner selection, Git collection, and authored path bases cannot observe
-		// different config-directory aliases within one document operation.
-		if s.fs != nil {
-			evaluationFS = newConfigSnapshotFS(bundled.WrapFS(cachedvfs.From(s.fs)))
-		}
-		if len(s.jsConfigs) > 0 {
-			jsOwnerIndex = target.NewOwnerIndex(s.jsConfigs, evaluationFS)
-		}
-	}
-	evaluateKnownConfig := func(
-		entries config.RslintConfig,
-		configDirectory string,
-		ownerIndex *target.OwnerIndex,
-		fileConfigResolver *config.FileConfigResolver,
-		catalog *rule.Catalog,
-		enforcePlugins bool,
-	) (config.RslintConfig, config.ResolvedFileConfig, *config.PathSpaceSnapshot, bool) {
-		configDirectory = tspath.NormalizePath(configDirectory)
-		if !s.configSnapshotIncludesGitignore {
-			entries = config.ConfigWithGitignoreForExactTarget(
-				entries,
-				configDirectory,
-				evaluationFS,
-				lintFile.Identity(),
-			)
-			ownerIndex = nil
-			fileConfigResolver = nil
-		}
-		if ownerIndex == nil {
-			ownerIndex = target.NewOwnerIndex(
-				map[string]config.RslintConfig{configDirectory: entries},
-				evaluationFS,
-			)
-		}
-		if _, ok := ownerIndex.PathSpaces().PhysicalDirectory(configDirectory); !ok {
-			return nil, config.ResolvedFileConfig{}, ownerIndex.PathSpaces(), false
-		}
-		if fileConfigResolver == nil {
-			if s.configSnapshotIncludesGitignore {
-				panic("committed config snapshot is missing its file config resolver")
-			}
-			var err error
-			fileConfigResolver, err = config.NewFileConfigResolverWithPathSpaces(
-				entries,
-				configDirectory,
-				evaluationFS,
-				ownerIndex.PathSpaces(),
-				catalog,
-				enforcePlugins,
-			)
-			if err != nil {
-				panic(err)
-			}
-		}
-		return entries, fileConfigResolver.ResolveTarget(lintFile.Identity()), ownerIndex.PathSpaces(), true
-	}
-
-	if len(s.jsConfigs) > 0 {
-		if s.configRefreshConfigPath != "" {
-			configKey := tspath.GetDirectoryPath(
-				tspath.NormalizePath(s.configRefreshConfigPath),
-			)
-			if entries, active := s.jsConfigs[configKey]; active {
-				entries, resolved, pathSpaces, ok := evaluateKnownConfig(
-					entries,
-					configKey,
-					jsOwnerIndex,
-					jsFileConfigResolvers[configKey],
-					jsRuleCatalog,
-					true,
-				)
-				return documentConfigSelection{
-					entries:       entries,
-					resolved:      resolved,
-					pathSpaces:    pathSpaces,
-					ruleCatalog:   jsRuleCatalog,
-					directory:     configKey,
-					configKey:     configKey,
-					usesJSConfig:  true,
-					configMissing: !ok,
-				}
-			}
-		} else if jsOwnerIndex != nil {
-			configKey, owned := jsOwnerIndex.Resolve(lintFile.Identity())
-			if owned {
-				entries := s.jsConfigs[configKey]
-				entries, resolved, pathSpaces, ok := evaluateKnownConfig(
-					entries,
-					configKey,
-					jsOwnerIndex,
-					jsFileConfigResolvers[configKey],
-					jsRuleCatalog,
-					true,
-				)
-				return documentConfigSelection{
-					entries:       entries,
-					resolved:      resolved,
-					pathSpaces:    pathSpaces,
-					ruleCatalog:   jsRuleCatalog,
-					directory:     configKey,
-					configKey:     configKey,
-					usesJSConfig:  true,
-					configMissing: !ok,
-				}
-			}
-		}
-	}
-
-	configDirectory := tspath.NormalizePath(s.cwd)
-	goRuleCatalog := rules.All()
-	entries, resolved, pathSpaces, ok := evaluateKnownConfig(
-		s.jsonConfig,
-		configDirectory,
-		s.jsonConfigOwnerIndex,
-		s.jsonFileConfigResolver,
-		goRuleCatalog,
-		false,
-	)
-	return documentConfigSelection{
-		entries:       entries,
-		resolved:      resolved,
-		pathSpaces:    pathSpaces,
-		ruleCatalog:   goRuleCatalog,
-		directory:     configDirectory,
-		configMissing: !ok,
-	}
-}
-
-// documentLintSnapshot resolves one immutable target identity, then derives
-// its owner, config, and TypeScript projects from that same identity. This is
-// the only production entry from an editor URI into lint configuration.
-func (s *Server) documentLintSnapshot(uri lsproto.DocumentUri) documentLintSnapshot {
-	target := lspTargetIdentity(uriToPath(uri), s.fs)
-	selection := s.selectDocumentConfig(target)
-	target.ConfigDirectory = selection.directory
-	typeScriptConfigPaths := s.tsConfigPaths
-	if selection.usesJSConfig {
-		typeScriptConfigPaths = s.tsConfigPathsByConfig[selection.configKey]
-	}
-	_, unavailable := s.jsUnavailableConfigs[selection.configKey]
-	return documentLintSnapshot{
-		target:                target,
-		config:                selection.entries,
-		resolvedConfig:        selection.resolved,
-		pathSpaces:            selection.pathSpaces,
-		ruleCatalog:           selection.ruleCatalog,
-		configResolved:        !selection.configMissing,
-		typeScriptConfigPaths: typeScriptConfigPaths,
-		usesJavaScriptConfig:  selection.usesJSConfig,
-		unavailable:           selection.usesJSConfig && unavailable,
-	}
-}
-
-// getConfigForURI is retained for package-level helpers and tests. Production
-// lint operations use documentLintSnapshot so config and project ownership
-// cannot be resolved in separate filesystem observations.
-func (s *Server) getConfigForURI(uri lsproto.DocumentUri) (config.RslintConfig, string, bool) {
-	target := lspTargetIdentity(uriToPath(uri), s.fs)
-	selection := s.selectDocumentConfig(target)
-	return selection.entries, selection.directory, selection.usesJSConfig
-}
-
-func (s *Server) getLintConfigForURI(uri lsproto.DocumentUri) (config.RslintConfig, string, bool) {
-	snapshot := s.documentLintSnapshot(uri)
-	return snapshot.config, snapshot.target.ConfigDirectory, snapshot.usesJavaScriptConfig
-}
-
-func (s *Server) isUnavailableConfigForURI(uri lsproto.DocumentUri) bool {
-	target := lspTargetIdentity(uriToPath(uri), s.fs)
-	configKey, ok := s.jsConfigKeyForTarget(target)
-	if !ok {
-		return false
-	}
-	_, unavailable := s.jsUnavailableConfigs[configKey]
-	return unavailable
-}
-
-// nearestJSConfigKey returns the nearest supplied JS/TS config directory for
-// uri. Matching uses normalized filesystem paths instead of URI identity;
-// lexical ownership is tried before a realpath fallback.
-func (s *Server) nearestJSConfigKey(uri lsproto.DocumentUri) (string, bool) {
-	return s.jsConfigKeyForTarget(lspTargetIdentity(uriToPath(uri), s.fs))
-}
-
-func (s *Server) jsConfigKeyForTarget(target target.File) (string, bool) {
-	if len(s.jsConfigs) == 0 {
-		return "", false
-	}
-	// An explicitly selected config governs every editor target, while its
-	// directory remains the base for files/ignores matching. Ancestry
-	// models automatic config discovery, not this invocation-wide scope.
-	if s.configRefreshConfigPath != "" {
-		configDir := tspath.GetDirectoryPath(tspath.NormalizePath(s.configRefreshConfigPath))
-		_, active := s.jsConfigs[configDir]
-		return configDir, active
-	}
-	if s.jsConfigOwnerIndex == nil {
-		return "", false
-	}
-	configDir, _ := s.jsConfigOwnerIndex.Resolve(target.Identity())
-	if configDir == "" {
-		return "", false
-	}
-	_, active := s.jsConfigs[configDir]
-	return configDir, active
-}
-
-// tsConfigPathsForURI returns parserOptions.project paths from the config owner
-// selected by getConfigForURI. A nested config with no tsconfig therefore does
-// not affect type-info decisions for sibling configs.
-//
-// A nil return means the governing config has no resolved tsconfig, so callers
-// must disable type-aware rules for this file.
-func (s *Server) tsConfigPathsForURI(uri lsproto.DocumentUri) []string {
-	target := lspTargetIdentity(uriToPath(uri), s.fs)
-	if configKey, ok := s.jsConfigKeyForTarget(target); ok {
-		return s.tsConfigPathsByConfig[configKey]
-	}
-	return s.tsConfigPaths
-}
-
-// pushDiagnostics runs the linter for the given URI and pushes results to the client.
-// Must be called synchronously from the LSP message loop (not from a goroutine)
-// because session is not goroutine-safe.
-func (s *Server) pushDiagnostics(uri lsproto.DocumentUri) {
-	if s.session == nil {
-		return
-	}
-
-	ctx := s.backgroundCtx
-
-	if !isLintableScriptFile(uri) {
-		return
-	}
-
-	// Supersede the previous plugin pass before linting. Native diagnostics are
-	// published synchronously below; the next plugin pass is stamped with this
-	// generation so the main loop can reject an older result.
-	s.docGeneration[uri]++
-	generation := s.docGeneration[uri]
-	s.cancelInflightPluginDispatch(uri)
-	snapshot := s.documentLintSnapshot(uri)
-	if snapshot.unavailable {
-		delete(s.diagnostics, uri)
-		if err := s.PublishDiagnostics(ctx, &lsproto.PublishDiagnosticsParams{
-			Uri:         uri,
-			Diagnostics: []*lsproto.Diagnostic{},
-		}); err != nil {
-			log.Printf("Error clearing diagnostics for unavailable config: %v", err)
-		}
-		return
-	}
-
-	lintResult, err := s.runConfiguredLint(uri, ctx, snapshot)
-	if err != nil {
-		log.Printf("Error running lint for push diagnostics: %v", err)
-		delete(s.diagnostics, uri)
-		if publishErr := s.PublishDiagnostics(ctx, &lsproto.PublishDiagnosticsParams{
-			Uri:         uri,
-			Diagnostics: []*lsproto.Diagnostic{},
-		}); publishErr != nil {
-			log.Printf("Error clearing diagnostics after lint failure: %v", publishErr)
-		}
-		return
-	}
-	ruleDiags := lintResult.Diagnostics
-
-	s.diagnostics[uri] = ruleDiags
-
-	// Must use empty slice (not nil) so JSON serializes as [] instead of null
-	lspDiags := make([]*lsproto.Diagnostic, 0, len(ruleDiags))
-	for _, d := range ruleDiags {
-		lspDiags = append(lspDiags, convertRuleDiagnosticToLSP(d))
-	}
-
-	if err := s.PublishDiagnostics(ctx, &lsproto.PublishDiagnosticsParams{
-		Uri:         uri,
-		Diagnostics: lspDiags,
-	}); err != nil {
-		log.Printf("Error publishing diagnostics: %v", err)
-	}
-
-	// Dispatch eslint-plugin rules off the main loop. The reverse request
-	// MUST NOT run synchronously here — it would block the dispatch loop (and
-	// thus all editor interaction) until the Node worker replies. Results merge
-	// back via pluginResultCh on the main loop (s.diagnostics is lock-free).
-	if !lintResult.HasSyntaxErrors {
-		s.dispatchPluginLintWithSnapshot(uri, generation, snapshot)
-	}
-
-	// The pacer cannot see that the lint just dropped what it derived.
-	go runtime.GC()
-}
